libc/fildes: Implement file offset in struct regfile_tx
Move file-offset handling from struct ofd_tx to struct file_tx.
